Implemeted indexed drawing

+++ include/si/vk_renderer.hpp
@@ -68,7 +68,7 @@ namespace si {
                 {{0.5f, 0.5f}, {0.0f, 1.0f, 0.0f}},
                 {{-0.5f, 0.5f}, {0.0f, 0.0f, 1.0f}}
             };
-            const std::vector<uint8_t> indices = {
+            const std::vector<uint16_t> indices = {
                 0, 1, 2                
             };
